Slide Images Left And Right
Created functions named moveLeft() and moveRight(). moveLeft() gets a
reference to the first slide, removes it from the parent element, and
appends it to the end of the parent element. moveRight gets a reference
to the last slide, removes it from the parent element, and inserts it
before the first slide, using .insertBefore() method.